agents/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "live-steam-agents"
version = "0.1.0"
description = "A2A + AG-UI agent backend for live-steam-app (orchestrator, component builder, and specialized agents)"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "Live Steam Team"}]
readme = "README.md"

dependencies = [
    # Google ADK for orchestrator agent
    "google-adk>=1.0.0",
    "google-genai>=0.3.0",
    
    # Agent frameworks
    "langgraph>=0.1.0",
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "langchain-google-genai>=0.1.0",
    
    # A2A Protocol
    "a2a-protocol>=1.0.0",
    
    # Web frameworks for agent servers
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.24.0",
    "starlette>=0.27.0",
    
    # Data validation and serialization
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # Environment management
    "python-dotenv>=1.0.0",
    
    # Async support
    "httpx>=0.24.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "ipython>=8.0.0",
]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
